AI-Psychologist/requirements.txt
streamlit>=1.28.0
streamlit-authenticator==0.4.2
PyYAML>=6.0
bcrypt>=4.0.0
requests>=2.28.0